+++ agent/check_handler.go
@@ -204,6 +204,8 @@ func extractMetrics(event *types.Event) []*types.MetricPoint {
 		transformer, err = transformers.ParseInflux(event.Check.Output)
 	case types.NagiosMetricFormat:
 		transformer, err = transformers.ParseNagios(event)
+	case types.OpenTSDBMetricFormat:
+		transformer, err = transformers.ParseOpenTSDB(event.Check.Output)
 	}
 
 	if err != nil {